+++ src/ui/gui/editor/widget_project_tabs/panel_classification.hpp
@@ -63,7 +63,7 @@ private:
   void populateClassesFromImage();
   void addClass(bool withUpdate = true);
   void createTableItem(int32_t rowIdx, enums::ClassId classId, const std::string &name, const std::string &color, const std::string &notes);
-  void moveClassToPosition(size_t fromPos, size_t newPos);
+  void moveClassToPosition(int32_t fromPos, int32_t newPos);
   auto findNextFreeClassId() -> enums::ClassId;
 
   /////////////////////////////////////////////////////
